feat: add queue listener
